See git://gist.github.com/538348.git for idea
Used
git clone git://gist.github.com/538348.git
to get the idea for this module.

Added safety measures to ensure that the code aborts at the first hint
of trouble.  Code will pick up where it left off, as soon as the
problem is repaired.  Code sets up a directory once again, to allow
the user to practice changing the code.